+++ modules/managed-node-group/main.tf
@@ -5,7 +5,7 @@ resource "aws_eks_node_group" "main" {
   subnet_ids             = var.node_group_subnet_ids
   ami_type               = var.ami_type
   capacity_type          = var.capacity_type
-  disk_size              = var.disk_size
+  disk_size              = var.create_custom_launch_template ? null : var.disk_size
   force_update_version   = var.force_update_version
   instance_types         = var.instance_types
   labels                 = var.labels
@@ -30,20 +30,10 @@ resource "aws_eks_node_group" "main" {
   }
 
   dynamic "launch_template" {
-    for_each = var.launch_template
+    for_each = var.create_custom_launch_template ? [1] : []
     content {
-      id      = launch_template.value.name != null ? null : lookup(launch_template.value, "id", null)
-      name    = launch_template.value.id != null ? null : lookup(launch_template.value, "name", null)
-      version = launch_template.value.version
-    }
-  }
-
-  dynamic "remote_access" {
-    for_each = var.remote_access
-    content {
-      #ec2_ssh_key = lookup(remote_access.value, "ec2_ssh_key", null)
-      ec2_ssh_key               = var.create_key_pair ? aws_key_pair.this[0].key_name : lookup(remote_access.value, "ec2_ssh_key", null)
-      source_security_group_ids = lookup(remote_access.value, "source_security_group_ids", [])
+      id      = local.launch_template_id
+      version = local.launch_template_version
     }
   }
 
@@ -60,28 +50,6 @@ resource "aws_eks_node_group" "main" {
     create = lookup(var.timeouts, "create", "60m")
     update = lookup(var.timeouts, "update", "60m")
     delete = lookup(var.timeouts, "delete", "60m")
-  }
-}
-
-### Key pair
-resource "tls_private_key" "this" {
-  count     = var.create_key_pair ? 1 : 0
-  algorithm = "RSA"
-  rsa_bits  = 4096
-}
-
-resource "aws_key_pair" "this" {
-  count      = var.create_key_pair ? 1 : 0
-  key_name   = "${var.cluster_name}-keypair"
-  public_key = tls_private_key.this[0].public_key_openssh
-  tags       = merge({ Name = "${var.cluster_name}-keypair" }, var.tags)
-}
-
-## For downloading the keypair to local computer
-resource "null_resource" "local_save_ec2_keypair" {
-  count = var.create_key_pair ? 1 : 0
-  provisioner "local-exec" {
-    command = "echo '${tls_private_key.this[0].private_key_pem}' > ${path.module}/${aws_key_pair.this[0].id}.pem"
   }
 }
 
@@ -114,4 +82,206 @@ resource "aws_iam_role_policy_attachment" "amazon_eks_cni_policy" {
 resource "aws_iam_role_policy_attachment" "amazon_ec2_container_registry_readonly" {
   policy_arn = "arn:aws:iam::aws:policy/AmazonEC2ContainerRegistryReadOnly"
   role       = aws_iam_role.node_group.name
+}
+
+## Managed Policy to allow ssm agent to communicate with SSM Manager
+resource "aws_iam_role_policy_attachment" "ssm" {
+  count      = var.create_custom_launch_template && var.install_ssm_agent ? 1 : 0
+  policy_arn = "arn:aws:iam::aws:policy/AmazonSSMManagedInstanceCore"
+  role       = aws_iam_role.node_group.name
+}
+
+resource "aws_launch_template" "main" {
+  #checkov:skip=CKV_AWS_79:Ensure Instance Metadata Service Version 1 is not enabled"
+  #checkov:skip=CKV_AWS_341:Ensure Launch template should not have a metadata response hop limit greater than 1
+  #checkov:skip=CKV_AWS_79:Ensure Instance Metadata Service Version 1 is not enabled
+  count                   = var.create_custom_launch_template ? 1 : 0
+  name                    = "${var.cluster_name}-custom-template"
+  description             = var.launch_template_description
+  ebs_optimized           = var.ebs_optimized
+  image_id                = var.image_id
+  instance_type           = var.instance_type
+  user_data               = var.install_ssm_agent ? data.template_cloudinit_config.config.rendered : var.user_data #use var.extra_script to run additional scripts
+  vpc_security_group_ids  = lookup(var.network_interfaces, "security_groups") != [] ? [] : var.security_group_ids
+  default_version         = var.default_version
+  disable_api_termination = var.disable_api_termination
+  kernel_id               = var.kernel_id
+  ram_disk_id             = var.ram_disk_id
+
+  dynamic "block_device_mappings" {
+    for_each = var.block_device_mappings
+    content {
+      device_name  = try(block_device_mappings.value.device_name, null)
+      no_device    = try(block_device_mappings.value.no_device, null)
+      virtual_name = try(block_device_mappings.value.virtual_name, null)
+
+      dynamic "ebs" {
+        for_each = flatten([try(block_device_mappings.value.ebs, [])])
+        content {
+          delete_on_termination = try(ebs.value.delete_on_termination, null)
+          encrypted             = try(ebs.value.encrypted, null)
+          kms_key_id            = try(ebs.value.kms_key_id, null)
+          iops                  = try(ebs.value.iops, null)
+          throughput            = try(ebs.value.throughput, null)
+          snapshot_id           = try(ebs.value.snapshot_id, null)
+          volume_size           = try(ebs.value.volume_size, null)
+          volume_type           = try(ebs.value.volume_type, null)
+        }
+      }
+    }
+  }
+
+  dynamic "capacity_reservation_specification" {
+    for_each = length(var.capacity_reservation_specification) > 0 ? [var.capacity_reservation_specification] : []
+    content {
+      capacity_reservation_preference = try(capacity_reservation_specification.value.capacity_reservation_preference, null)
+
+      dynamic "capacity_reservation_target" {
+        for_each = try([capacity_reservation_specification.value.capacity_reservation_target], [])
+        content {
+          capacity_reservation_id                 = try(capacity_reservation_target.value.capacity_reservation_id, null)
+          capacity_reservation_resource_group_arn = try(capacity_reservation_target.value.capacity_reservation_resource_group_arn, null)
+        }
+      }
+    }
+  }
+
+  dynamic "cpu_options" {
+    for_each = length(var.cpu_options) > 0 ? [var.cpu_options] : []
+    content {
+      core_count       = cpu_options.value.core_count
+      threads_per_core = cpu_options.value.threads_per_core
+    }
+  }
+
+  dynamic "credit_specification" {
+    for_each = length(var.credit_specification) > 0 ? [var.credit_specification] : []
+    content {
+      cpu_credits = credit_specification.value.cpu_credits
+    }
+  }
+
+  dynamic "elastic_gpu_specifications" {
+    for_each = length(var.elastic_gpu_specifications) > 0 ? [var.elastic_gpu_specifications] : []
+    content {
+      type = elastic_gpu_specifications.value.type
+    }
+  }
+
+  dynamic "elastic_inference_accelerator" {
+    for_each = length(var.elastic_inference_accelerator) > 0 ? [var.elastic_inference_accelerator] : []
+    content {
+      type = elastic_inference_accelerator.value.type
+    }
+  }
+
+  dynamic "enclave_options" {
+    for_each = length(var.enclave_options) > 0 ? [var.enclave_options] : []
+    content {
+      enabled = enclave_options.value.enabled
+    }
+  }
+
+  dynamic "instance_market_options" {
+    for_each = length(var.instance_market_options) > 0 ? [var.instance_market_options] : []
+    content {
+      market_type = instance_market_options.value.market_type
+
+      dynamic "spot_options" {
+        for_each = try([instance_market_options.value.spot_options], [])
+        content {
+          block_duration_minutes         = try(spot_options.value.block_duration_minutes, null)
+          instance_interruption_behavior = try(spot_options.value.instance_interruption_behavior, null)
+          max_price                      = try(spot_options.value.max_price, null)
+          spot_instance_type             = try(spot_options.value.spot_instance_type, null)
+          valid_until                    = try(spot_options.value.valid_until, null)
+        }
+      }
+    }
+  }
+
+  dynamic "license_specification" {
+    for_each = length(var.license_specifications) > 0 ? [var.license_specifications] : []
+    content {
+      license_configuration_arn = license_specifications.value.license_configuration_arn
+    }
+  }
+
+  dynamic "metadata_options" {
+    for_each = length(keys(var.metadata_options)) > 0 ? [var.metadata_options] : []
+    content {
+      http_endpoint               = try(metadata_options.value.http_endpoint, null)
+      http_protocol_ipv6          = try(metadata_options.value.http_protocol_ipv6, null)
+      http_put_response_hop_limit = try(metadata_options.value.http_put_response_hop_limit, null)
+      http_tokens                 = try(metadata_options.value.http_tokens, null)
+      instance_metadata_tags      = try(metadata_options.value.instance_metadata_tags, null)
+    }
+  }
+
+  monitoring {
+    enabled = var.enable_monitoring
+  }
+
+  dynamic "network_interfaces" {
+    for_each = var.network_interfaces
+    content {
+      associate_carrier_ip_address = try(network_interfaces.value.associate_carrier_ip_address, null)
+      associate_public_ip_address  = try(network_interfaces.value.associate_public_ip_address, null)
+      delete_on_termination        = try(network_interfaces.value.delete_on_termination, false)
+      description                  = try(network_interfaces.value.description, null)
+      device_index                 = try(network_interfaces.value.device_index, null)
+      interface_type               = try(network_interfaces.value.interface_type, null)
+      ipv4_prefix_count            = try(network_interfaces.value.ipv4_prefixes, null) != null ? null : try(network_interfaces.value.ipv4_prefix_count, null)
+      ipv4_prefixes                = try(network_interfaces.value.ipv4_prefixes, null)
+      ipv4_addresses               = try(network_interfaces.value.ipv4_addresses, null)
+      ipv4_address_count           = try(network_interfaces.value.ipv4_addresses, null) != null ? null : try(network_interfaces.value.ipv4_address_count, null)
+      ipv6_prefix_count            = try(network_interfaces.value.ipv6_prefixes, []) != [] ? null : try(network_interfaces.value.ipv6_prefix_count, null)
+      ipv6_prefixes                = try(network_interfaces.value.ipv6_prefixes, [])
+      ipv6_addresses               = try(network_interfaces.value.ipv6_addresses, null)
+      ipv6_address_count           = try(network_interfaces.value.ipv6_addresses, null) != null ? null : try(network_interfaces.value.ipv6_address_count, null)
+      network_interface_id         = try(network_interfaces.value.network_interface_id, null)
+      network_card_index           = try(network_interfaces.value.network_card_index, null)
+      private_ip_address           = try(network_interfaces.value.private_ip_address, null)
+      security_groups              = compact(concat(try(network_interfaces.value.security_groups, []), var.security_group_ids))
+    }
+  }
+
+  dynamic "placement" {
+    for_each = length(var.placement) > 0 ? [var.placement] : []
+    content {
+      affinity                = lookup(placement.value, "affinity", null)
+      availability_zone       = lookup(placement.value, "availability_zone", null)
+      group_name              = lookup(placement.value, "group_name", null)
+      host_id                 = lookup(placement.value, "host_id", null)
+      host_resource_group_arn = lookup(placement.value, "host_resource_group_arn", null)
+      spread_domain           = lookup(placement.value, "spread_domain", null)
+      tenancy                 = lookup(placement.value, "tenancy", null)
+      partition_number        = lookup(placement.value, "partition_number", null)
+    }
+  }
+
+  dynamic "private_dns_name_options" {
+    for_each = length(var.private_dns_name_options) > 0 ? [var.private_dns_name_options] : []
+    content {
+      enable_resource_name_dns_aaaa_record = lookup(private_dns_name_options.value, "enable_resource_name_dns_aaaa_record", null)
+      enable_resource_name_dns_a_record    = lookup(private_dns_name_options.value, "enable_resource_name_dns_a_record", null)
+      hostname_type                        = private_dns_name_options.value.hostname_type
+    }
+  }
+
+  dynamic "tag_specifications" {
+    for_each = var.tag_specifications
+    content {
+      resource_type = try(tag_specifications.value.resource_type, "volume")
+      tags          = var.tags
+    }
+  }
+
+  lifecycle {
+    create_before_destroy = true
+  }
+
+  depends_on = [
+    aws_iam_role.node_group
+  ]
 }